Experiments/video_narration/requirements (1).txt
openai
google-cloud-vision
google-auth
google-generativeai
diffusers
torch
streamlit
transformers
accelerate
moviepy
SentencePiece
elevenlabs
Transformers